Extract loader output to `getLoaderOutput`
